Reorganize into workspace, split core and chains implementation
