+++ Cargo.toml
@@ -1,22 +1,6 @@
-[package]
-name = "farcaster_core"
-version = "0.1.0"
-authors = ["h4sh3d <h4sh3d@protonmail.com>"]
-edition = "2018"
+[workspace]
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-curve25519-dalek = "3.0.0"
-hex = "0.4.3"
-rand_core = { version = "^0.5.0", features = ["getrandom"] }
-secp256k1 = { version = "0.20.1", features = ["rand-std"] }
-strict_encoding = "1.2.1"
-thiserror = "1.0.24"
-
-# blockchain specific
-bitcoin = "0.26.0"
-monero = { version = "0.11.0", features = ["strict_encoding_support"] }
-
-[dev-dependencies]
-bitcoincore-rpc = "0.13.0"
+members = [
+    "core",
+    "chains",
+]
