+++ chains/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "farcaster_chains"
+version = "0.1.0"
+authors = ["h4sh3d <h4sh3d@protonmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+farcaster_core = { path = "../core" }
+strict_encoding = "1.2.1"
+hex = "0.4.3"
+
+# blockchain specific
+bitcoin = "0.26.0"
+monero = { version = "0.11.0", features = ["strict_encoding_support"] }
+
+[dev-dependencies]
+bitcoincore-rpc = "0.13.0"
+rand_core = { version = "^0.5.0", features = ["getrandom"] }
+secp256k1 = { version = "0.20.1", features = ["rand-std"] }
