+++ chains/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod bitcoin;
+pub mod monero;
