+++ core/src/datum.rs
