+++ core/src/instruction.rs
