+++ core/src/lib.rs
@@ -3,14 +3,12 @@
 #[macro_use]
 pub mod consensus;
 
-pub mod bitcoin;
 pub mod blockchain;
 pub mod bundle;
 pub mod crypto;
 pub mod datum;
 pub mod instruction;
 pub mod interactions;
-pub mod monero;
 pub mod negotiation;
 pub mod protocol_message;
 pub mod role;
