+++ core/src/version.rs
